% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating_assignment.R
\name{rating_assignment}
\alias{rating_assignment}
\title{Randomly assign items to raters, so that each item is assigned to pair_size raters}
\usage{
rating_assignment(items, raters, pair_size)
}
\arguments{
\item{items}{A vector of items that need to be rated}

\item{raters}{A vector of rater identifiers}

\item{pair_size}{The number of raters that must rate each item}
}
\value{
A list of 3 data frames. goal_record shows the names of raters
  assigned to each goal. rater_record shows all the goals assigned to each
  rater. sum_record shows a summary including the total number of goals being
  assigned, and the total number of goals assigned to each rater.
}
\description{
Randomly assign items to raters, so that each item is assigned to pair_size raters
}
